requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
numpy==2.3.1
pandas==2.3.0
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
